+++ zcash_client_backend/src/data_api/wallet.rs
@@ -8,17 +8,10 @@ use zcash_primitives::{
         components::{amount::DEFAULT_FEE, Amount},
         Transaction,
     },
-    zip32::{AccountId, ExtendedFullViewingKey, ExtendedSpendingKey},
 };
 
 #[cfg(feature = "transparent-inputs")]
-use {
-    zcash_address::unified::Typecode,
-    zcash_primitives::{
-        legacy::keys::{self as transparent, IncomingViewingKey},
-        sapling::keys::OutgoingViewingKey,
-    },
-};
+use zcash_primitives::{legacy::keys::IncomingViewingKey, sapling::keys::OutgoingViewingKey};
 
 use crate::{
     address::RecipientAddress,
@@ -27,6 +20,7 @@ use crate::{
         SentTransactionOutput, WalletWrite,
     },
     decrypt_transaction,
+    keys::UnifiedSpendingKey,
     wallet::OvkPolicy,
     zip321::{Payment, TransactionRequest},
 };
@@ -129,7 +123,7 @@ where
 /// };
 /// use zcash_proofs::prover::LocalTxProver;
 /// use zcash_client_backend::{
-///     keys::sapling,
+///     keys::UnifiedSpendingKey,
 ///     data_api::{wallet::create_spend_to_address, error::Error, testing},
 ///     wallet::OvkPolicy,
 /// };
@@ -148,8 +142,8 @@ where
 /// };
 ///
 /// let account = AccountId::from(0);
-/// let extsk = sapling::spending_key(&[0; 32][..], COIN_TYPE, account);
-/// let to = extsk.default_address().1.into();
+/// let usk = UnifiedSpendingKey::from_seed(&Network::TestNetwork, &[0; 32][..], account).unwrap();
+/// let to = usk.to_unified_full_viewing_key().default_address().0.into();
 ///
 /// let mut db_read = testing::MockWalletDb {
 ///     network: Network::TestNetwork
@@ -159,8 +153,7 @@ where
 ///     &mut db_read,
 ///     &Network::TestNetwork,
 ///     tx_prover,
-///     account,
-///     &extsk,
+///     &usk,
 ///     &to,
 ///     Amount::from_u64(1).unwrap(),
 ///     None,
@@ -176,8 +169,7 @@ pub fn create_spend_to_address<E, N, P, D, R>(
     wallet_db: &mut D,
     params: &P,
     prover: impl TxProver,
-    account: AccountId,
-    extsk: &ExtendedSpendingKey,
+    usk: &UnifiedSpendingKey,
     to: &RecipientAddress,
     amount: Amount,
     memo: Option<MemoBytes>,
@@ -206,8 +198,7 @@ where
         wallet_db,
         params,
         prover,
-        extsk,
-        account,
+        usk,
         &req,
         ovk_policy,
         min_confirmations,
@@ -248,7 +239,7 @@ where
 /// * `wallet_db`: A read/write reference to the wallet database
 /// * `params`: Consensus parameters
 /// * `prover`: The TxProver to use in constructing the shielded transaction.
-/// * `extsk`: The extended spending key that controls the funds that will be spent
+/// * `usk`: The unified spending key that controls the funds that will be spent
 ///   in the resulting transaction.
 /// * `account`: The ZIP32 account identifier associated with the extended spending
 ///   key that controls the funds to be used in creating this transaction.  This
@@ -265,8 +256,7 @@ pub fn spend<E, N, P, D, R>(
     wallet_db: &mut D,
     params: &P,
     prover: impl TxProver,
-    extsk: &ExtendedSpendingKey,
-    account: AccountId,
+    usk: &UnifiedSpendingKey,
     request: &TransactionRequest,
     ovk_policy: OvkPolicy,
     min_confirmations: u32,
@@ -277,12 +267,11 @@ where
     R: Copy + Debug,
     D: WalletWrite<Error = E, TxRef = R>,
 {
-    // Check that the ExtendedSpendingKey we have been given corresponds to the
-    // ExtendedFullViewingKey for the account we are spending from.
-    let extfvk = ExtendedFullViewingKey::from(extsk);
-    if !wallet_db.is_valid_account_extfvk(account, &extfvk)? {
-        return Err(E::from(Error::InvalidExtSk(account)));
-    }
+    let account = wallet_db
+        .get_account_for_ufvk(&usk.to_unified_full_viewing_key())?
+        .ok_or(Error::KeyNotRecognized)?;
+
+    let extfvk = usk.sapling().to_extended_full_viewing_key();
 
     // Apply the outgoing viewing key policy.
     let ovk = match ovk_policy {
@@ -335,7 +324,12 @@ where
         let merkle_path = selected.witness.path().expect("the tree is not empty");
 
         builder
-            .add_sapling_spend(extsk.clone(), selected.diversifier, note, merkle_path)
+            .add_sapling_spend(
+                usk.sapling().clone(),
+                selected.diversifier,
+                note,
+                merkle_path,
+            )
             .map_err(Error::Builder)?;
     }
 
@@ -452,8 +446,7 @@ pub fn shield_transparent_funds<E, N, P, D, R, U>(
     wallet_db: &mut D,
     params: &P,
     prover: impl TxProver,
-    sk: &transparent::AccountPrivKey,
-    account: AccountId,
+    usk: &UnifiedSpendingKey,
     memo: &MemoBytes,
     min_confirmations: u32,
 ) -> Result<D::TxRef, E>
@@ -463,28 +456,20 @@ where
     R: Copy + Debug,
     D: WalletWrite<Error = E, TxRef = R> + WalletWriteTransparent<UtxoRef = U>,
 {
-    // Obtain the UFVK for the specified account & use its internal change address
-    // as the destination for shielded funds.
-    let shielding_address = wallet_db
-        .get_unified_full_viewing_keys()
-        .and_then(|ufvks| {
-            ufvks
-                .get(&account)
-                .ok_or_else(|| E::from(Error::AccountNotFound(account)))
-                .and_then(|ufvk| {
-                    // TODO: select the most preferred shielded receiver once we have the ability to
-                    // spend Orchard funds.
-                    ufvk.sapling()
-                        .map(|dfvk| dfvk.change_address().1)
-                        .ok_or_else(|| E::from(Error::KeyNotFound(account, Typecode::Sapling)))
-                })
-        })?;
+    let account = wallet_db
+        .get_account_for_ufvk(&usk.to_unified_full_viewing_key())?
+        .ok_or(Error::KeyNotRecognized)?;
 
+    let shielding_address = usk
+        .sapling()
+        .to_diversifiable_full_viewing_key()
+        .change_address()
+        .1;
     let (latest_scanned_height, latest_anchor) = wallet_db
         .get_target_and_anchor_heights(min_confirmations)
         .and_then(|x| x.ok_or_else(|| Error::ScanRequired.into()))?;
 
-    let account_pubkey = sk.to_account_pubkey();
+    let account_pubkey = usk.transparent().to_account_pubkey();
     let ovk = OutgoingViewingKey(account_pubkey.internal_ovk().as_bytes());
 
     // derive the t-address for the extpubkey at the minimum valid child index
@@ -510,7 +495,10 @@ where
 
     let mut builder = Builder::new_with_fee(params.clone(), latest_scanned_height, fee);
 
-    let secret_key = sk.derive_external_secret_key(child_index).unwrap();
+    let secret_key = usk
+        .transparent()
+        .derive_external_secret_key(child_index)
+        .unwrap();
     for utxo in &utxos {
         builder
             .add_transparent_input(secret_key, utxo.outpoint.clone(), utxo.txout.clone())